pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flickr-to-google-photo"
version = "0.1.0"
description = "Migrate photos from Flickr to Google Photos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flickrapi>=2.4",
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "requests>=2.31",
    "click>=8.1",
    "python-dotenv>=1.0",
    "piexif>=1.1.3",
]

[project.scripts]
flickr-to-gphoto = "flickr_to_google_photo.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
]

[tool.hatch.build.targets.wheel]
packages = ["src/flickr_to_google_photo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100
